app/src/main/java/sokolov/dunkancards/home/presenter/CategoriesPresenterImpl.java
package sokolov.dunkancards.home.presenter;

import sokolov.dunkancards.home.interactor.CategoriesInteractor;
import sokolov.dunkancards.home.view.CategoriesView;

public class CategoriesPresenterImpl implements CategoriesPresenter {

    private final CategoriesView categoriesView;
    private final CategoriesInteractor categoriesInteractor;

    public CategoriesPresenterImpl(CategoriesView categoriesView, CategoriesInteractor categoriesInteractor) {
        this.categoriesView = categoriesView;
        this.categoriesInteractor = categoriesInteractor;
    }

    @Override
    public void onViewShow() {
        categoriesView.setCategories(
                categoriesInteractor.loadCategories());
    }
}
